// src/redux/user/user.types.ts
enum UserTypes {
	SIGN_UP_USER_START = "SIGN_UP_USER_START",
	SIGN_UP_USER_SUCCESS = "SIGN_UP_USER_SUCCESS",
	SIGN_UP_USER_FAILURE = "SIGN_UP_USER_FAILURE",
	SIGN_IN_USER_START = "SIGN_IN_USER_START",
	SIGN_IN_USER_SUCCESS = "SIGN_IN_USER_SUCCESS",
	SIGN_IN_USER_FAILURE = "SIGN_IN_USER_FAILURE",
	SET_USER_AUTH = "SET_USER_AUTH",
	CHECK_USER_SESSION = "CHECK_USER_SESSION",
	SIGN_OUT_USER = "SIGN_OUT_USER",
	SET_PROFILE_PICTURE = "SET_PROFILE_PICTURE",
	SEARCH_USERS_START = "SEARCH_USERS_START",
	SEARCH_USERS_SUCCESS = "SEARCH_USERS_SUCCESS",
	CLEAR_USER_SEARCH = "CLEAR_USER_SEARCH",
	SET_SEARCH_PAGINATION = "SET_SEARCH_PAGINATION",
}

export default UserTypes;